+++ src/Machinery/DisposableStateMachine_2.cs
@@ -14,9 +14,6 @@ namespace Machinery
 
         public DisposableStateMachine(TContext context, IState<TContext, TEvent> initialState)
         {
-            if (context is null)
-                throw new ArgumentNullException(nameof(context));
-
             _context = context;
             _currentState = initialState ?? throw new ArgumentNullException(nameof(initialState));
         }