+++ src/Machinery/StateMachine_3.cs
@@ -25,9 +25,6 @@ namespace Machinery
 
         public StateMachine(TContext context, TState initialState)
         {
-            if (context is null)
-                throw new ArgumentNullException(nameof(context));
-
             if (initialState is null)
                 throw new ArgumentNullException(nameof(initialState));