+++ src/Machinery/StateMachine_4.cs
@@ -26,9 +26,6 @@ namespace Machinery
 
         public StateMachine(TContext context, TState initialState, TPolicy policy)
         {
-            if (context is null)
-                throw new ArgumentNullException(nameof(context));
-
             if (initialState is null)
                 throw new ArgumentNullException(nameof(initialState));