+++ src/Machinery/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.4.0] - 2019-10-20
+### Added
+- `IDisposable` constraint for `TState` type parameter of `DisposableStateMachine<…>` generics.
+
+### Removed
+- Context constraint to be nonnull.
+
 ## [0.3.0] - 2019-10-16
 ### Added
 - `DisposableStateMachine<…>` classes owning their states.
@@ -22,6 +29,7 @@
 ### Added
 - Basic building blocks for generic state machines: `StateMachine<…>`, `IState<…>`, `IPolicy<…>`.
 
+[0.4.0]: https://github.com/qbit86/machinery/compare/machinery-0.3.0...machinery-0.4.0
 [0.3.0]: https://github.com/qbit86/machinery/compare/machinery-0.2.0...machinery-0.3.0
 [0.2.0]: https://github.com/qbit86/machinery/compare/machinery-0.1.1...machinery-0.2.0
 [0.1.1]: https://github.com/qbit86/machinery/compare/machinery-0.1.0...machinery-0.1.1